Add robot-arm host provider and operation-aware resource write flow
### Motivation
- Introduce a mock `robot-arm` host provider and example to exercise host command semantics and integrate it into the CLI host registration flow.
- Extend the runtime resource model to be operation-aware so context send targets can map to custom capability names (not just legacy `read`/`write`).
- Validate and canonicalize host manifest operations and enable providers to receive operation metadata during preflight/write.

### Description
- Add a new host crate `mech-host-robot-arm` with a manifest (`host.mcfg`), provider implementation, and a small example under `examples/robot-arm-demo` to demonstrate commands and state reads.
- Propagate an `operation: RuntimeCapabilityOperation` field through `RuntimeResourceWritePreflightRequest` and `RuntimeResourceWriteRequest` and adjust provider APIs and implementations (CLI, browser, in-memory docs, tests) to pass the operation when preflighting and writing.
- Add runtime support for operation-aware capability checks by introducing `host_interface::operation` utilities and updating host manifest validation and `preflight_context_access`/write logic to resolve the correct operation for `Assign` vs `Send` intents (including `context_send_operation`, `context_write_operation`, and related helpers).
- Add `src/cli/host_factories.rs` to register built-in host factories (CLI and conditional `robot-arm`) and update `new_cli_runtime` to use the registry; update `RuntimeHostFactoryRegistry` with helper methods for provider discovery.
- Update and extend tests and add new runtime tests for custom operations/plotter behavior and robot provider behavior; update many existing tests to account for the new `operation` field.

### Testing
- Ran unit tests for runtime and hosts with `cargo test` across the workspace; all existing and new tests for resource preflight, host providers, and runtime operation-dispatch passed.
- Updated provider compatibility tests for `browser` and `cli` to include `RuntimeCapabilityOperation` and they succeed.
- Added unit tests for `RobotArmResourceProvider` and new plotter/operation behavior in `src/runtime/tests` and they succeed when running the test suite.
